pyproject.toml
[project]
name = "memora"
version = "0.2.0"
description = "Inteligência Técnica Operacional — assistente de suporte baseado no codebase"
requires-python = ">=3.11"
dependencies = [
    "fastapi~=0.115",
    "uvicorn[standard]~=0.34",
    "pydantic-settings~=2.7",
    "sqlalchemy~=2.0",
    "psycopg2-binary~=2.9",
    "pgvector~=0.3",
    "httpx~=0.28",
    "anthropic~=0.43",
    "openai~=1.60",
    "tiktoken~=0.7",
    "python-dotenv~=1.0",
    "gitpython~=3.1",
    "python-jose[cryptography]~=3.3",
    "cryptography~=44.0",
    "groq~=0.5",
    "google-generativeai~=0.8",
    "python-multipart~=0.0.18",
    "bcrypt~=4.0",
    "pypdf~=5.0",
    "python-docx~=1.1",
    "xhtml2pdf~=0.2.16",
    "markdown~=3.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.25",
    "ruff>=0.9",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
